requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
oauthlib==3.3.1
pandas==2.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
